Style Remaining

+++ backend/controllers/userController.js
@@ -16,14 +16,8 @@ const createToken = (_id) => {
 // login user
 const loginUser = async (req, res) => {
     const {username, password} = req.body;
-    const emptyFields = [];
-
-    if(!username){
-        emptyFields.push('username');
-    }
-    if(!password){
-        emptyFields.push('password');
-    }
+    let emptyFields = [];
+    emptyFields = checkEmptyFields({username, password})
 
     if(emptyFields.length > 0){
         return res.status(400).json({error: 'Fill out all blanks', emptyFields});
@@ -49,29 +43,16 @@ const loginUser = async (req, res) => {
     const gender = user.gender;
     const privilege = user.privilege;
     const _id = user._id;
-    const imageurl = user.imageurl;
+    const imageUrl = user.imageUrl;
 
-    res.status(200).json({username, fullname, email, contact, address, description, gender, privilege, imageurl, _id, token});
+    res.status(200).json({username, fullname, email, contact, address, description, gender, privilege, imageUrl, _id, token});
 };
 
 // client signup route
 const signupClient = async (req, res) => {
     const {username, fullname, email, password} = req.body;
-
-    const emptyFields = [];
-
-    if(!username){
-        emptyFields.push('username');
-    }
-    if(!password){
-        emptyFields.push('password');
-    }
-    if(!fullname){
-        emptyFields.push('fullname');
-    }
-    if(!email){
-        emptyFields.push('email');
-    }
+    let emptyFields = [];
+    emptyFields = checkEmptyFields({username, fullname, email, password})
 
     // validation
     if(emptyFields.length > 0){
@@ -109,76 +90,64 @@ const signupClient = async (req, res) => {
 const updateProfile = async (req, res) => {
     upload.single('image')(req, res, async err => {
         if(err){
-            return res.status(400).json({error: err.message});
+            return res.status(400).json({error: err.message, emptyFields: []});
         }
         const {username, fullname, email, contact, address, description, gender, changePWD, oldPWD, newPWD, removePic} = req.body;
-        const emptyFields = [];
-
-        if(!username){
-            emptyFields.push('username');
-        }
-        if(!fullname){
-            emptyFields.push('fullname');
-        }
-        if(!email){
-            emptyFields.push('email');
-        }
-        if(!contact){
-            emptyFields.push('contact');
-        }
-        if(!gender){
-            emptyFields.push('gender');
-        }
-        if(changePWD){
-            if(!oldPWD){
-                emptyFields.push('oldPWD');
-            }
-            if(!newPWD){
-                emptyFields.push('newPWD');
-            }
-        }
+        let emptyFields = [];
+        emptyFields = checkEmptyFields({username, fullname, email, contact, gender});
+        emptyFields = changePWD.toString() === 'true' ? 
+            checkEmptyFields({username, fullname, email, contact, gender, oldPWD, newPWD}) : 
+            checkEmptyFields({username, fullname, email, contact, gender});
 
         // validation
         if(emptyFields.length > 0){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Fill out all blanks', emptyFields});
         }
         if(!validator.isEmail(email)){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Invalid email format', emptyFields});
         }
         if(newPWD && !validator.isStrongPassword(newPWD)){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Weak password. The password should contain UPPER case, lower case, number and symbol keys.', emptyFields});
         }
 
         const exists = await User.findOne({username});
         
         if(exists && exists._id.toString() !== req.user._id.toString()){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Username already in use', emptyFields});
         }
 
         const exists1 = await User.findOne({email});
         
         if(exists1 && exists1._id.toString() !== req.user._id.toString()){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Email already in use', emptyFields});
         }
 
         const exists2 = await User.findOne({contact});
         
         if(exists2 && exists2._id.toString() !== req.user._id.toString()){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'Phone Number already in use', emptyFields});
         }
 
         const exists3 = await User.findOne({fullname, privilege: 'admin'});
         
         if(exists3 && req.user.privilege === 'admin' && exists3._id.toString() !== req.user._id.toString()){
+            req.file && fs.unlink(req.file.path, imageDeleteError);
             return res.status(400).json({error: 'RealEstate name already in use', emptyFields});
         }
 
         let hash;
-        if(changePWD){
+        if(changePWD.toString() === 'true'){
             const user = await User.findById(req.user._id).select('password');
 
             const match = await bcrypt.compare(oldPWD, user.password);
             if(!match){
+                req.file && fs.unlink(req.file.path, imageDeleteError);
                 return res.status(400).json({error: 'Incorrect Password', emptyFields});
             }
             const salt = await bcrypt.genSalt(10);
@@ -187,21 +156,16 @@ const updateProfile = async (req, res) => {
 
         let imageUrl;
         if(req.file){
-            imageUrl = req.file.path
+            imageUrl = req.file.path;
         }
-        if(removePic){
+        if(removePic.toString() === 'true'){
             imageUrl = '';
         }
-        const updatedUser = await User.findByIdAndUpdate(req.user._id, {username, fullname, email, contact, address, description, gender, password: hash, imageUrl });
+
+        const updatedUser = await User.findByIdAndUpdate(req.user._id, {username, fullname, email, contact, address, description, gender, password: hash, imageUrl});
         const oldImageUrl = updatedUser.imageUrl;
-        if(oldImageUrl && (removePic || imageUrl) && fs.existsSync(oldImageUrl)){
-            fs.unlink(oldImageUrl, err => {
-                if(err){
-                    console.log(`Error deleting Image: ${err}`);
-                } else{
-                    console.log('Image deleted successfully');
-                }
-            });
+        if(oldImageUrl && (removePic.toString() === 'true' || imageUrl) && fs.existsSync(oldImageUrl)){
+            fs.unlink(oldImageUrl, imageDeleteError);
         } else{
             console.log('No Image Deleted');
         }
@@ -216,22 +180,8 @@ const updateProfile = async (req, res) => {
 // admin signup route
 const signupAdmin = async (req, res) => {
     const {username, fullname, email, contact, password, privilege='admin'} = req.body;
-    const emptyFields = [];
-    if(!username){
-        emptyFields.push('username');
-    }
-    if(!fullname){
-        emptyFields.push('fullname');
-    }
-    if(!contact){
-        emptyFields.push('contact');
-    }
-    if(!email){
-        emptyFields.push('email');
-    }
-    if(!password){
-        emptyFields.push('password');
-    }
+    let emptyFields = [];
+    emptyFields = checkEmptyFields({username, fullname, email, contact, password});
 
     // validation
     if(emptyFields.length > 0){
@@ -263,12 +213,17 @@ const signupAdmin = async (req, res) => {
     const salt = await bcrypt.genSalt(10);
     const hash = await bcrypt.hash(password, salt);
 
-    const admin = await User.create({username, fullname, contact, email, password: hash, privilege});
+    const admin = (await User.create({username, fullname, contact, email, password: hash, privilege}));
 
+    const cr_at = admin.createdAt;
+    const id= admin._id;
     const fn = admin.fullname;
     const pvg = admin.privilege;
+    const ctc = admin.contact;
+    const eml = admin.email;
 
-    res.status(200).json({fullname: fn, privilege: pvg});
+
+    res.status(200).json({_id: id, fullname: fn, privilege: pvg, email: eml, contact: ctc, createdAt: cr_at});
 }
 
 // admin delete route
@@ -279,31 +234,22 @@ const deleteAdmin = async (req, res) => {
     }
 
     const user = await User.findById(id);
+    if(!user){
+        return res.status(400).json({error: 'No admin found'});
+    }
     if(user.privilege !== 'admin'){
         return res.status(400).json({error: 'You can only delete an admin'});
     }
 
     const admin = await User.findByIdAndDelete(id);
-    if(!admin){
-        return res.status(400).json({error: 'No admin found'});
-    }
+    
 
     let apartments = await Apartment.find({realestate_id: id}).select('imageUrls');
     if(apartments && apartments.length > 0){
         apartments.forEach(apartment => {
             const imageUrls = apartment.imageUrls;
             if(imageUrls && imageUrls > 0){
-                imageUrls.forEach(imageUrl => {
-                    if(fs.existsSync(imageUrl)){
-                        fs.unlink(imageUrl, err => {
-                            if(err){
-                                console.log(`Error deleting Image: ${err}`);
-                            } else{
-                                console.log('Image deleted successfully');
-                            }
-                        });
-                    }
-                });
+                imageUrls.forEach(imageUrl => fs.existsSync(imageUrl) && fs.unlink(imageUrl, imageDeleteError));
             } else{
                 console.log('No Image Deleted');
             }
@@ -311,24 +257,16 @@ const deleteAdmin = async (req, res) => {
     }
 
     apartments = await Apartment.deleteMany({realestate_id: id});
-    const request = await Request.updateMany({realestate_id: id}, {realestate_name: 'Sorry, This Apartment has been removed', status: 'pending'});
+    const request = await Request.updateMany({realestate_id: id}, {
+        realestate_name: 'Sorry, This Apartment has been removed', 
+        status: 'pending'
+    });
     
     if(admin.imageUrl && fs.existsSync(admin.imageUrl)){
-        fs.unlink(admin.imageUrl, err => {
-            if(err){
-                console.log(`Error deleting Image: ${err}`);
-            } else{
-                console.log('Image deleted successfully');
-            }
-        });
-    } else{
-        console.log('No Image Deleted');
+        fs.unlink(admin.imageUrl, imageDeleteError);
     }
 
-    const fn = admin.fullname;
-    const pvg = admin.privilege;
-
-    res.status(200).json({fullname: fn, privilege: pvg}, apartments, request);
+    res.status(200).json({_id: admin._id});
 }
 
 // get admin route
@@ -337,6 +275,19 @@ const getAdmins = async (req, res) => {
     const admins = await User.find({privilege}).select('-password').sort({createdAt: -1});
 
     res.status(200).json(admins);
+}
+
+// commonly used functions
+const checkEmptyFields = elements => {
+    let emptyFields = []
+    for(let key in elements){
+        const value = elements[key];
+        !value && emptyFields.push(key);
+    }
+    return emptyFields;
+}
+const imageDeleteError = err => {
+    err ? console.log(`Error deleting Image: ${err}`) : console.log('Image deleted successfully');
 }
 
 module.exports = { loginUser, signupClient, updateProfile, signupAdmin, deleteAdmin, getAdmins };
